requirements.txt
akshare>=1.11.10
pandas>=1.3.0
sqlalchemy>=1.4.0
python-dotenv>=0.19.0
schedule>=1.1.0
pymysql>=1.0.2
loguru>=0.6.0 